refactor: move svgr plugin and generate ts script to root
