+++ packages/icons/package.json
@@ -26,7 +26,7 @@
     "build:cjs": "tsup --outDir dist/cjs --format=cjs",
     "build:esm": "tsup --format=esm --legacy-output",
     "build:native": "tsup src/**/*.svg --outDir dist/cjs/svg --format=cjs -- native",
-    "build:types": "node scripts/generate-d-ts.js src/**/*.ts",
+    "build:types": "node ../../scripts/generate-d-ts.js src/**/*.ts",
     "build:types:native": "yarn build:types --native",
     "prepublishOnly": "node ../../scripts/prepublish.js"
   },
