+++ scripts/generate-d-ts.js
